Refactor subcommand package
